Seems to be working broken out voting from content-item
